# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "E-Commerce Demo API",
    "version": "1.0.0",
    "description": "花卉電商網站 REST API"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "sessionId": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Session-Id"
      }
    }
  },
  "paths": {
...
    "/api/ecpay/checkout/{orderId}": {
      "post": {
        "summary": "建立 ECPay 綠界付款交易，回傳 HTML auto-submit form",
        "tags": [
          "ECPay"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功，回傳 HTML form"
          },
          "400": {
            "description": "訂單狀態不是 pending 或已有付款記錄"
          },
          "404": {
            "description": "訂單不存在"
          }
        }
      }
    },
    "/api/ecpay/status/{orderId}": {
      "get": {
        "summary": "主動查詢 ECPay 付款狀態（QueryTradeInfo）",
        "tags": [
          "ECPay"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查詢成功"
          },
          "400": {
            "description": "尚未建立 ECPay 交易"
          },
          "404": {
            "description": "訂單不存在"
          }
        }
      }
    },
...
  },
...
}